src/handler/requirements.txt
boto3>=1.35.0
pyjwt[crypto]>=2.9.0
pyyaml>=6.0.0
